Use a variable format for schema users
Added new variable mon_mysql_users and mon_influxdb_users that are arrays with
entries with username and password attributes. This allows ansible variables to be used
as user names.

The old mysql_users and influxdb_users are left in place to allow for a phased transition
to the new variables. Once all uses of those variables have been changed to the new ones,
the old ones will be deleted